+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - Refactor IntelliJ Platform type resolution and validation to exclude unavailable distribution constrains JetBrains/intellij-platform-gradle-plugin#2027
+- Refactor `ide` and `ides` methods to rely on the cacheable IntelliJ Platform dependency creation.
 
 ## [2.9.0] - 2025-09-07